Moved the SocialLinksView to other modules
