+++ Packages/PrimitivesComponents/Sources/Types/InsightLink.swift
@@ -1,5 +1,4 @@
 // Copyright (c). Gem Wallet. All rights reserved.
-
 import Foundation
 import SwiftUI
